Implement sub-millisecond UUID ordering
This patch adds a new behavior to the UUID generator such that on a
single machine, if more than one UUID is generated in the same wallclock
millisecond, the UUIDs will sort in the order they were created. This is
useful for test code that generates fragments quickly and wants to
ensure that the fragments are ordered by generation time.